+++ libs/application/workspace/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/workspace-application-surface.service';
 export * from './lib/workspace-back.service';
+export * from './lib/workspace-navigation.models';
 export * from './lib/workspace-surface.models';
 export * from './lib/workspace-search.models';